requirements.txt
click              >=8.0.0,  <9.0.0
rich               >=14.0.0, <15.0.0
pysimdjson         >=7.0.0,  <8.0.0